requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
aiobotocore==2.4.2
    # via s3fs
aiohttp==3.8.3
    # via
    #   aiobotocore
    #   s3fs
aioitertools==0.11.0
    # via aiobotocore
aiosignal==1.3.1
    # via aiohttp
asn1crypto==1.5.1
    # via
    #   oscrypto
    #   scramp
    #   snowflake-connector-python
async-timeout==4.0.2
    # via aiohttp
attrs==22.2.0
    # via aiohttp
botocore==1.27.59
    # via aiobotocore
certifi==2022.12.7
    # via
    #   requests
    #   snowflake-connector-python
cffi==1.15.1
    # via
    #   cryptography
    #   snowflake-connector-python
charset-normalizer==2.1.1
    # via
    #   aiohttp
    #   requests
    #   snowflake-connector-python
cryptography==38.0.4
    # via
    #   pyopenssl
    #   snowflake-connector-python
filelock==3.9.0
    # via snowflake-connector-python
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
fsspec==2022.11.0
    # via s3fs
greenlet==2.0.1
    # via sqlalchemy
idna==3.4
    # via
    #   requests
    #   snowflake-connector-python
    #   yarl
jinja2==3.1.2
    # via sqla-raw
jmespath==1.0.1
    # via botocore
markupsafe==2.1.1
    # via jinja2
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
oscrypto==1.3.0
    # via snowflake-connector-python
pg8000==1.29.4
    # via -r requirements.in
psycopg2-binary==2.9.5
    # via -r requirements.in
pycparser==2.21
    # via cffi
pycryptodomex==3.16.0
    # via snowflake-connector-python
pyjwt==2.6.0
    # via snowflake-connector-python
pyopenssl==22.1.0
    # via snowflake-connector-python
python-dateutil==2.8.2
    # via
    #   botocore
    #   pg8000
python-dotenv==0.21.0
    # via -r requirements.in
pytz==2022.7.1
    # via snowflake-connector-python
requests==2.28.2
    # via snowflake-connector-python
s3fs==2022.11.0
    # via sqla-raw
scramp==1.4.4
    # via pg8000
six==1.16.0
    # via python-dateutil
snowflake-connector-python==2.9.0
    # via snowflake-sqlalchemy
snowflake-sqlalchemy==1.4.4
    # via -r requirements.in
sqla-raw==1.5.0
    # via -r requirements.in
sqlalchemy==1.4.46
    # via
    #   snowflake-sqlalchemy
    #   sqla-raw
typing-extensions==4.4.0
    # via
    #   aioitertools
    #   snowflake-connector-python
urllib3==1.26.14
    # via
    #   botocore
    #   requests
    #   snowflake-connector-python
wrapt==1.14.1
    # via aiobotocore
yarl==1.8.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
